requirements.txt
streamlit as st
argparse
time
pyaudio
wave
pyttsx3
speech_recognition
googletrans
langchain.vectorstores.chroma
langchain.prompts
langchain_community.llms.ollama
loader
gtts
os
tempfile
playsound
